+++ docs/_data/tree.json
@@ -20,7 +20,7 @@
     },
     {
       "tree": "sanddance-specs",
-      "url": "/docs/sanddance-specs/v2"
+      "url": "/docs/sanddance-specs/v1"
     },
     {
       "tree": "sanddance-react",
