+++ docs/examples/index.md
@@ -8,7 +8,7 @@ title: Examples
 ## Subsystem
 
 ### sanddance-specs
-* [sanddance-specs 2D vega specs tests](../tests/sanddance-specs/v2/) using UMD/CDN
+* [sanddance-specs 2D vega specs tests](../tests/sanddance-specs/v1/) using UMD/CDN
 
 ### cube-layer
 * [cubeTest](../tests/v3/umd/cubeTest.html) using UMD/CDN
