+++ docs/tests/sanddance-specs/v1/img/usa-counties-equirectangular.png
