+++ docs/tests/sanddance-specs/v1/index.html
@@ -50,6 +50,10 @@
         <label>
             Insight file:
             <select id="select-spec">
+                <option>scatter-with-image.json</option>
+                <option>scatter.json</option>
+                <option>scatter-facet-wrap.json</option>
+                <option>scatter-facet-wrap-with-image.json</option>
                 <option>column-categoric.json</option>
                 <option>column-facet-cross.json</option>
                 <option>column-facet-wrap.json</option>
@@ -62,11 +66,11 @@
                 <option>density-treemap.json</option>
                 <option>density.json</option>
                 <option>scatter-facet-cross.json</option>
-                <option>scatter-facet-wrap.json</option>
-                <option>scatter.json</option>
                 <option>stacks-facet-cross.json</option>
                 <option>stacks-facet-wrap.json</option>
-                <option>stacks-not-faceted.json</option>
+                <option>stacks.json</option>
+                <option>stacks-with-image.json</option>
+                <option>stacks-facet-wrap-with-image.json</option>
             </select>
         </label>
     </header>
