+++ docs/tests/sanddance-specs/v1/specs/scatter-facet-wrap-with-image.json
