+++ docs/tests/sanddance-specs/v1/specs/scatter-with-image.json
