+++ docs/tests/sanddance-specs/v1/specs/stacks-facet-wrap-with-image.json
