+++ docs/tests/sanddance-specs/v1/specs/stacks-with-image.json
