--- docs/tests/sanddance-specs/v2/specs/column-facet-cross.json
@@ -1,17 +1,0 @@
-{
-    "columns": {
-        "x": "Income",
-        "color": "Obama",
-        "z": "Education",
-        "facet": "MedAge",
-        "facetV": "Income"
-    },
-    "facetStyle": "cross",
-    "scheme": "redblue",
-    "size": {
-        "height": 600,
-        "width": 800
-    },
-    "chart": "barchartV",
-    "view": "3d"
-}
