--- docs/tests/sanddance-specs/v2/specs/column-facet-wrap.json
@@ -1,16 +1,0 @@
-{
-    "columns": {
-        "x": "Income",
-        "color": "Obama",
-        "z": "Education",
-        "facet": "MedAge"
-    },
-    "facetStyle": "wrap",
-    "scheme": "redblue",
-    "size": {
-        "height": 600,
-        "width": 800
-    },
-    "chart": "barchartV",
-    "view": "3d"
-}
