--- docs/tests/sanddance-specs/v2/specs/column-sum-strip.json
@@ -1,16 +1,0 @@
-{
-    "columns": {
-        "x": "Income",
-        "color": "Obama",
-        "z": "Education",
-        "size": "TotalPop"
-    },
-    "totalStyle": "sum-strip",
-    "scheme": "redblue",
-    "size": {
-        "height": 600,
-        "width": 800
-    },
-    "chart": "barchartV",
-    "view": "3d"
-}
