--- docs/tests/sanddance-specs/v2/specs/column.json
@@ -1,14 +1,0 @@
-{
-    "columns": {
-        "x": "Income",
-        "color": "Obama",
-        "z": "Education"
-    },
-    "scheme": "redblue",
-    "size": {
-        "height": 600,
-        "width": 800
-    },
-    "chart": "barchartV",
-    "view": "3d"
-}
