--- docs/tests/sanddance-specs/v2/specs/density-facet-wrap.json
@@ -1,17 +1,0 @@
-{
-    "columns": {
-        "x": "Longitude",
-        "y": "Latitude",
-        "color": "Obama",
-        "z": "Education",
-        "facet": "MedAge"
-    },
-    "facetStyle": "wrap",
-    "scheme": "redblue",
-    "size": {
-        "height": 600,
-        "width": 800
-    },
-    "chart": "density",
-    "view": "3d"
-}
