--- docs/tests/sanddance-specs/v2/specs/density.json
@@ -1,15 +1,0 @@
-{
-    "columns": {
-        "x": "Longitude",
-        "y": "Latitude",
-        "color": "Obama",
-        "z": "Education"
-    },
-    "scheme": "redblue",
-    "size": {
-        "height": 600,
-        "width": 800
-    },
-    "chart": "density",
-    "view": "3d"
-}
