--- docs/tests/sanddance-specs/v2/specs/scatter-facet-cross.json
@@ -1,18 +1,0 @@
-{
-    "columns": {
-        "x": "Longitude",
-        "y": "Latitude",
-        "color": "Obama",
-        "z": "Education",
-        "facet": "MedAge",
-        "facetV": "Income"
-    },
-    "facetStyle": "cross",
-    "scheme": "redblue",
-    "size": {
-        "height": 600,
-        "width": 800
-    },
-    "chart": "scatterplot",
-    "view": "3d"
-}
