--- docs/tests/sanddance-specs/v2/specs/stacks-facet-cross.json
@@ -1,19 +1,0 @@
-{
-    "columns": {
-        "x": "Longitude",
-        "y": "Latitude",
-        "color": "Obama",
-        "z": "Education",
-        "sort": "State",
-        "facet": "MedAge",
-        "facetV": "Income"
-    },
-    "facetStyle": "cross",
-    "scheme": "redblue",
-    "size": {
-        "height": 600,
-        "width": 800
-    },
-    "chart": "stacks",
-    "view": "3d"
-}
