--- docs/tests/sanddance-specs/v2/specs/stacks-facet-wrap.json
@@ -1,18 +1,0 @@
-{
-    "columns": {
-        "x": "Longitude",
-        "y": "Latitude",
-        "color": "Obama",
-        "z": "Education",
-        "sort": "State",
-        "facet": "MedAge"
-    },
-    "facetStyle": "wrap",
-    "scheme": "redblue",
-    "size": {
-        "height": 600,
-        "width": 800
-    },
-    "chart": "stacks",
-    "view": "3d"
-}
