--- docs/tests/sanddance-specs/v2/specs/stacks.json
@@ -1,16 +1,0 @@
-{
-    "columns": {
-        "x": "Longitude",
-        "y": "Latitude",
-        "color": "Obama",
-        "z": "Education",
-        "sort": "State"
-    },
-    "scheme": "redblue",
-    "size": {
-        "height": 600,
-        "width": 800
-    },
-    "chart": "stacks",
-    "view": "3d"
-}
