+++ packages/sanddance-specs/README.md
@@ -60,9 +60,8 @@ You will need an array of data objects, so that columns types can be inferred.
 
 ## Versions
 
-### 2.0.0 Changes
+### 1.5.0 Changes
 
-* BREAKING: `Size` interface removed, moved to @msrvida/chart-types
 * Added background image
 
 ### 1.4.0 Changes
